requirements.txt
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.9.0
buildhat==0.7.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorzero==2.0
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
distro==1.9.0
fonttools==4.57.0
gpiozero==2.0.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jmespath==0.10.0
kiwisolver==1.4.8
matplotlib==3.10.1
nls @ file:///Users/doskey/src/alibabacloud-nls-python-sdk
numpy==2.2.4
openai==1.72.0
opencv-python==4.11.0.86
oss2==2.19.1
packaging==24.2
pillow==11.1.0
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
